Leaderboard refactor
Fixes #1757
